+++ packages/gdu/utils/configs.ts
@@ -5,3 +5,9 @@ export const getConfigsDirs = () =>
 		findUp.sync('.gdu_config', { type: 'directory' }),
 		findUp.sync('.gdu_app_config', { type: 'directory' }),
 	].filter(Boolean);
+
+export const getBuildFolder = () =>
+	[
+		findUp.sync('.gdu_config', { type: 'directory' }),
+		findUp.sync('.gdu_app_config', { type: 'directory' }),
+	].filter(Boolean);